Make JSON test helper public
Move helper to support engine, include it in pageflow and scrolled
engines via public helper.
