+++ spec/support/helpers/public_json_helper.rb
@@ -1,0 +1,6 @@
+require 'pageflow/json_test_helper'
+
+RSpec.configure do |config|
+  config.include(Pageflow::JsonTestHelper, type: :controller)
+  config.include(Pageflow::JsonTestHelper, type: :helper)
+end
